refactor: 모달 내부 컴포넌트 분리
