+++ apps/client/src/shared/component/NewFileImportModal/NewFileImportModal.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import { IcFileUpload } from '@tiki/icon';
-import { Button, Flex, Text, scrollStyle, useToastAction } from '@tiki/ui';
+import { Flex, scrollStyle, useToastAction } from '@tiki/ui';
 
 import { useEffect, useState } from 'react';
 
@@ -9,14 +8,10 @@ import { useQueryClient } from '@tanstack/react-query';
 import { $api } from '@/shared/api/client';
 import { Files } from '@/shared/api/time-blocks/team/time-block/type';
 import { Modal } from '@/shared/component/Modal';
-import {
-  boxStyle,
-  scrollBoxStyle,
-  uploadBoxStyle,
-} from '@/shared/component/NewFileImportModal/NewFileImportModal.style';
+import FileUploadContainer from '@/shared/component/NewFileImportModal/FileUploadContainer/FileUploadContainer';
+import { scrollBoxStyle } from '@/shared/component/NewFileImportModal/NewFileImportModal.style';
 import { DocumentDetail } from '@/shared/component/TimeBlockModal';
 import { flexStyle } from '@/shared/component/TimeBlockModal/component/UploadModal/UploadModal.style';
-import useFile from '@/shared/component/TimeBlockModal/hook/common/useFile';
 import UploadedFileItem from '@/shared/component/UploadedFileItem/UploadedFileItem';
 import { useInitializeTeamId } from '@/shared/hook/common/useInitializeTeamId';
 import { useCloseModal, useModalIsOpen } from '@/shared/store/modal';
@@ -29,7 +24,7 @@ interface NewFileImportModalProps {
   onNext?: () => void;
 }
 
-interface FileWithDocumentId extends File {
+export interface FileWithDocumentId extends File {
   documentId?: number;
 }
 
@@ -54,21 +49,6 @@ const NewFileImportModal = ({
   const { mutate: postDocumentMutation } = $api.useMutation('post', '/api/v1/teams/{teamId}/documents');
   const { mutate: deleteDocumentMutation } = $api.useMutation('delete', '/api/v1/teams/{teamId}/documents');
   const queryClient = useQueryClient();
-
-  const { fileInputRef, handleFileChange, handleDragOver, handleDrop } = useFile({
-    files,
-    onFilesChange: (newFiles) => {
-      setFiles((prevFiles) => {
-        const uniqueFiles = newFiles.filter(
-          (newFile) => !prevFiles.some((file) => file.name === newFile.name && file.size === newFile.size)
-        );
-
-        return [...prevFiles, ...uniqueFiles];
-      });
-    },
-    setFileUrls,
-    setUploadStatus,
-  });
 
   const handleUploadFile = (validFiles: FileWithDocumentId[]) => {
     postDocumentMutation(
@@ -171,35 +151,13 @@ const NewFileImportModal = ({
       <Modal.Body>
         <Flex css={flexStyle}>
           <Flex styles={{ direction: 'column', width: '100%' }}>
-            <Flex
-              css={[boxStyle, uploadBoxStyle(size)]}
-              onDragOver={handleDragOver}
-              onDrop={(event) => handleDrop(event)}>
-              <Flex tag="form" styles={{ direction: 'column', align: 'center', justify: 'center' }}>
-                <input
-                  type="file"
-                  multiple
-                  style={{ display: 'none' }}
-                  ref={fileInputRef}
-                  onChange={handleFileChange}
-                />
-                <IcFileUpload width={32} height={32} css={{ marginBottom: '1.2rem' }} />
-                <Flex styles={{ direction: 'column', gap: '0.8rem', align: 'center', justify: 'center' }}>
-                  <Text tag="body6" css={{ marginBottom: '0.8rem' }}>
-                    업로드할 파일을 끌어다 놓으세요.
-                  </Text>
-                  <Text tag="body8" css={{ whiteSpace: 'nowrap' }}>
-                    JPEG, PNG, PDF, Word 형식의 파일을 업로드할 수 있습니다.
-                  </Text>
-                  <Text tag="body8" css={{ marginBottom: '2rem' }}>
-                    최대 파일 크기는 50MB입니다.
-                  </Text>
-                </Flex>
-                <Button variant="outline" css={{ width: '16rem' }} onClick={() => fileInputRef.current?.click()}>
-                  파일 브라우저
-                </Button>
-              </Flex>
-            </Flex>
+            <FileUploadContainer
+              size={size}
+              files={files}
+              setFiles={setFiles}
+              setFileUrls={setFileUrls}
+              setUploadStatus={setUploadStatus}
+            />
             <div css={[scrollStyle, scrollBoxStyle(size)]}>
               {files.map((file) => (
                 <UploadedFileItem
